Renames new icons
