+++ packages/mobile-top-nav/src/assets/img/icon-ellipses.js
