+++ packages/mobile-top-nav/src/assets/img/icon-ia-logo.js
