+++ packages/mobile-top-nav/src/assets/img/icon-images.js
