+++ packages/mobile-top-nav/src/assets/img/icon-texts.js
