+++ packages/mobile-top-nav/src/assets/img/icon-video.js
