+++ packages/mobile-top-nav/src/assets/img/icon-web.js
