+++ packages/mobile-top-nav/src/assets/img/icons.js
@@ -1,0 +1,19 @@
+import audio from './icon-audio';
+import ellipses from './icon-ellipses';
+import iaLogo from './icon-ia-logo';
+import images from './icon-images';
+import software from './icon-software';
+import texts from './icon-texts';
+import video from './icon-video';
+import web from './icon-web';
+
+export default {
+  audio,
+  ellipses,
+  iaLogo,
+  images,
+  software,
+  texts,
+  video,
+  web,
+};
