--- packages/mobile-top-nav/src/assets/img/static_icons.js
@@ -1,19 +1,0 @@
-import audio from './icon-audio-static';
-import ellipses from './icon-ellipses-static';
-import iaLogo from './icon-ia-logo-static';
-import images from './icon-images-static';
-import software from './icon-software-static';
-import texts from './icon-texts-static';
-import video from './icon-video-static';
-import web from './icon-web-static';
-
-export default {
-  audio,
-  ellipses,
-  iaLogo,
-  images,
-  software,
-  texts,
-  video,
-  web,
-};
